Update it.h
Hi guy. Today I had the good idea to compare the english FMVs of TR1 with the french version, in order to understand better some slangs. 
I've edited again some parts of the italian translation in order to make it sharper (and more correct, of course). Now it should be perfect! :)